+++ dvr-scan.py
@@ -11,7 +11,7 @@
 # to using `python -m dvr_scan`) to prevent confusion with installed copies.
 # Programmatically, this file is identical to the dvr_scan/__main__.py file.
 #
-# Copyright (C) 2016-2021 Brandon Castellano <http://www.bcastell.com>.
+# Copyright (C) 2016-2022 Brandon Castellano <http://www.bcastell.com>.
 #
 # DVR-Scan is licensed under the BSD 2-Clause License; see the included
 # LICENSE file or visit one of the following pages for details: