+++ docs/changelog.md
@@ -11,6 +11,7 @@ DVR-Scan Changelog
  * [bugfix] Allow a maximum window size to be set when using `-roi` ([#59](https://github.com/Breakthrough/DVR-Scan/issues/59)):
      * The `-roi` flag now accepts a maximum window size for the ROI selection window (e.g. `-roi 1920 1080`)
      * If the `screeninfo` package is installed, or you are using a Windows build, videos will automatically be resized to the maximum screen size
+ * [bugfix] Fix frozen timestamp at beginning of video ([#68](https://github.com/Breakthrough/DVR-Scan/issues/68), was frozen for time specified by `-tb`/`--time-before`)
 
 #### Known Issues